Object Pooler Created
It's a pretty user-friendly object pooler, in the end. It'll work for
anything that derives from Component. The only gotchas are that you are
given a Handle to keep track of instead of the object itself (the handle
holds the reference to the object, though) and the handle is needed to
return the object. Other that than that, though, it works well as a drop
in replacement for Instantiate/Destroy. Doesn't work with GameObjects,
though.

The backend code is still a mess though. It needs A LOT of tidying up.
The solution is pretty elegant, but all still in one file, with some
unused pieces. I'll do that next