Clean up ObjectPoolsManager
Added some regions to improve readability and removed some superfluous
methods. Also added some quality of access methods. Removed direct
access to the component reference from Handle, so that accesses to
invalid objects is harder. UnmanagedDeref can still access the object if
need, provided the object is valid at time of call. Preferably, though,
a call to CallIfValid is used to access the object, so that calls only
ever act on valid objects

note: internally, an object is refered to as both "invalid" and "not
active" when it's in the pool ready to be borrowed. "Valid" and equally
"active" objects are the ones that have been borrowed from the pool